requirements.txt
openenv-core>=0.2.3
fastapi
uvicorn
sentence-transformers
bandit
radon
psutil
httpx
openai
pydantic
torch
# (Dependencies extracted from pyproject.toml)
